verifica califiacion

+++ src/Controller/StudyControlController.php
@@ -1169,14 +1169,23 @@ class StudyControlController extends AbstractController {
 			$id_inscription = $request->request->get('id_inscription');
 			$identity = $jwtauth->checkToken($token, true);
 			$inscription =  $em->getRepository(Inscription::class)->findOneById($id_inscription);
-			$helpers->binnacleAction('Inscription','elimino',$createdAt,'Se retiro el estudiante id='.$id_inscription,$identity->id);
-			$em->remove($inscription);
-			$em->flush();
-			$response = array(
-				'status' => 'success',
-				'code' => 200,
-				'msg' => 'Estudiante retirado.',
-			);
+			$qualifi =  $em->getRepository(Qualification::class)->findOneById($inscription);
+			if($qualifi) {
+				$response = array(
+					'status' => 'error',
+					'code' => 400,
+					'msg' => 'No se puede retirar porque posee notas.',
+				);
+			} else {
+				$helpers->binnacleAction('Inscription','elimino',$createdAt,'Se retiro el estudiante id='.$id_inscription,$identity->id);
+				$em->remove($inscription);
+				$em->flush();
+				$response = array(
+					'status' => 'success',
+					'code' => 200,
+					'msg' => 'Estudiante retirado.',
+				);
+			}
 		} else {
 			$response = array(
 				'status' => 'error',
